add program config page
