+++ client/src/store/constant.js
@@ -80,6 +80,8 @@ const DEMO = {
     return `/profile/${user_id}`;
   },
   PROGRAMS: '/programs',
+  KEYWORDS_EDIT: '/courses/analysis/keywords',
+  CREATE_NEW_PROGRAM_ANALYSIS: '/courses/analysis/programs/new',
   SINGLE_PROGRAM_LINK: (program_id) => {
     return `/programs/${program_id}`;
   },
